refactor: padronize returns and change id type

+++ internal/product/repository.go
@@ -3,11 +3,11 @@ package product
 import "github.com/igorverse/go-web-server-poc/internal/domain"
 
 type Repository interface {
-	Get(id uint64) (domain.Product, error)
+	Get(id int) (domain.Product, error)
 	GetAll() ([]domain.Product, error)
 	Store(p domain.Product) (domain.Product, error)
 	Update(p domain.Product) (domain.Product, error)
-	lastID() (uint64, error)
+	lastID() (int, error)
 }
 
 type repository struct{}
